Align naming with ES terminology
This commit renames some cocepts to be better aligned with ES names. For example, the nodes are now called n1,n2,n3. The concensus oracle is called master. Also, I re-arranged the nodes in the output to have the C,n1,n2,n3 and the primary to be n1 (and future primaries to have a lower order). This gives better graphs.